config: Allow DNS changes
Before, HAproxy resolved hostnames once at startup, and kept the same
association for the lifetime of the service, even if the IP changed in
DNS. This is obviously incorrect when load balanced containers get
restarted with new IPs. HAproxy now tries to updates IPs when health
checks to the old IP fail.